Added a decorator that supports multiple hypothesis examples.
